Exercise 9 do while loop with false condition
This prints "Hello World!" once then terminates. The difference between
the while and do while is the former is a pre-condition test and the
latter is a post-condition test. For a post-condition test the content
of the loop will always be executed at least once before the termination
condition is checked.